Add links to all reports under the course > reports menu
We found it very useful for teachers that are used to see reports under the course>reports menu, to find those (beautiful and useful) reports without the need to add an instance of the block to their course.
Please see if it looks useful to you too.
btw, we are not using Hoppot, so it was not part of this patch. be advised.